+++ src/common/js/index.js
@@ -3,11 +3,10 @@
  */
 
 import './global.js';
-//
 import './constants.js';
 import './i18n.js';
-//
 import './configure.js';
 //
-import './functions.js';
+import './checks.js';
+import './preferred.js';
 import './startup.js';